add(LCI_lbuffer_alloc/free): allocate and free long buffer. refactor lci.c

+++ src/include/server/server.h
@@ -20,7 +20,7 @@ typedef uint32_t LCID_meta_t; // immediate data
   lc_pool* pkpool; \
   struct lc_rep* rep;  \
   LCI_mt_t mt; \
-  size_t recv_posted; \
+  int recv_posted; \
   uintptr_t heap_addr; \
   LCID_mr_t heap_mr;
